Like Me - patch notes 0.6.6
Сделал(а) рефакторинг кода.
- Перенесла логику регистрации пользователя в отдельный файл  user_registration.py
- Переименовала имена функций связанных с регистрацией в main для лучшей читаемости кода.
- Проверила, что ничего не сломалось.
